Only log explicitly logged WIL failures in release builds
The WIL failure handler was logging failures when throwing an exception or returning a failure result, as well as when the failure was being explicitly logged.

This caused the same failure to be logged multiple times.

This changes the logic so that failures are only logged in release builds if they were being explicitly logged.