Check services are available in LowLevelMouseHookManager event handler
There was an odd bug-check crash in `LowLevelMouseHookManager::HookThread::s_on_event()` when foobar2000 was shutting down. Unfortunately the crash was from a foobar2000 preview version for which symbols are no longer available. Best guess, however, is it the hook was somehow called after services had shut down. This therefore adds a check that services are available before queuing a main thread callback.